+++ examples/custom_metrics.py
@@ -7,19 +7,21 @@ from bytewax.inputs import (
     DynamicSource,
     StatelessSourcePartition,
 )
-from bytewax.metrics import Gauge
+from prometheus_client import Gauge
 
 
 class PeriodicPartition(StatelessSourcePartition):
-    def __init__(self, frequency: timedelta):
+    def __init__(self, step_id: str, frequency: timedelta):
         self.frequency = frequency
         self._next_awake = datetime.now(timezone.utc)
         self._gauge = Gauge(
             "next_batch_delay",
-            "Calculated delay of when next batch was called",
-            ["partition"],
+            "Calculated delay of when next batch was called in seconds",
+            ["step_id", "partition"],
+            unit="seconds",
         )
         self._counter = 0
+        self._metric_labels = {"step_id": step_id, "partition": "0"}
 
     def next_awake(self):
         return self._next_awake
@@ -30,7 +32,7 @@ class PeriodicPartition(StatelessSourcePartition):
         # to  be called, and when it is actually called
         delay = datetime.now(timezone.utc) - self._next_awake
         self._next_awake += self.frequency
-        self._gauge.set_val(delay.total_seconds() * 1000, {"partition": "0"})
+        self._gauge.labels(**self._metric_labels).set(delay.total_seconds())
 
         return [self._counter]
 
@@ -39,8 +41,8 @@ class PeriodicSource(DynamicSource):
     def __init__(self, frequency):
         self.frequency = frequency
 
-    def build(self, worker_index: int, worker_count: int):
-        return PeriodicPartition(frequency=self.frequency)
+    def build(self, step_id: str, worker_index: int, worker_count: int):
+        return PeriodicPartition(step_id, frequency=self.frequency)
 
 
 flow = Dataflow("custom_metrics")